refactor: self-contained Facebook post content
For most online Zambians, social media *is* the internet.
A single ecosystem, led by Facebook, commands the vast majority
of user attention and online activity. Facebook and WhatsApp
form an integrated ecosystem that is the default for
communication, news, and business discovery.
This dominance is reinforced by "social bundles" from
mobile operators, which offer cheap access to these platforms.